Day 8
This algorithm is a little bit slow (to be fair, it's basically just the most naive version of it). As I was finishing, I remembered an algorithm that might be similar. I then looked it up; Prim's algorithm is *similar* to what I want; it or something similar could probably be adapted. (As I now need sleep, this is left as an exercise to the reader.)